Crate solana_example_mocks

Mock types for use in examples.

These represent APIs from crates that themselves depend on this crate, and which are useful for illustrating the examples for APIs in this crate.

Directly depending on these crates though would cause problematic circular dependencies, so instead they are mocked out here in a way that allows examples to appear to use crates that this crate must not depend on.

Each mod here has the name of a crate, so that examples can be structured to appear to import from that crate.